Loading target distance, run on health from monster.xml Use monsters .xml files to configure the strategies

+++ mtanksl.OpenTibia.FileFormats/Xml/Monsters/ChangeTargetStrategy.cs
@@ -8,6 +8,6 @@ namespace OpenTibia.FileFormats.Xml.Monsters
         public int Interval { get; set; }
 
         [XmlAttribute("chance")]
-        public int Chance { get; set; }
+        public double Chance { get; set; }
     }
 }
