+++ mtanksl.OpenTibia.FileFormats/Xml/Monsters/TargetStrategy.cs
@@ -5,15 +5,15 @@ namespace OpenTibia.FileFormats.Xml.Monsters
     public class TargetStrategy
     {
         [XmlAttribute("nearest")]
-        public int Nearest { get; set; }
+        public double Nearest { get; set; }
 
         [XmlAttribute("weakest")]
-        public int Weakest { get; set; }
+        public double Weakest { get; set; }
 
         [XmlAttribute("mostdamaged")]
-        public int MostDamaged { get; set; }
+        public double MostDamaged { get; set; }
 
         [XmlAttribute("random")]
-        public int Random { get; set; }
+        public double Random { get; set; }
     }
 }
